add psram support, with a huge caveat
This is a translation of the example over in
https://github.com/Altaflux/rp2350-psram-test to embassy, but I've been
scratching my head wondering why it didn't work in the picocalc. The
reason is that the psram is on the picocalc carrier board rather than
the microcontroller itself, and the mcu chipselect lines are not broken
out, so it cannot possibly work that way.

I'm committing this for posterity, but a subsequent commit will
have to talk spi directly and use some kind of block based IO or DMA
to actually work with the chip.