logging: write to both own cdc and via host serial
Duplicate log streams to both the usb cdc logger setup directly
on the pico, as well as the uart that is routed via the host
picocalc board and via a usb-to-serial chip.

The latter is convenient because it allows the usb-serial connection
to persist across a reset of the pico, eliminating an early startup
race condition where you are rushing to re-establish the usb
connection directly on the pico.