introduce Process trait and LocalShell
We now divert keyboard input to an instance of LocalShell,
which accumulates a command and "executes" it when you hit
enter.

Execution is just printing what you wrote, before then
clearing the command buffer and waiting for more input.

Serial input is synthesized into KeyReports and directed
along the same path.